pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "text2video-dfloat11"
version = "0.1.0"
description = "text2video-dfloat11"
readme = "README.md"
requires-python = "=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers",
    "fastapi>=0.123.0",
    "peft>=0.18.0",
    "platformdirs>=4.0.0",
    "python-multipart>=0.0.20",
    "sdnq",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
    "mcp>=1.23.2",
]

[project.scripts]
text2video = "text2video.cli:main"


[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers" }
sdnq = { git = "https://github.com/Disty0/sdnq" }

[tool.hatch.build.targets.wheel]
packages = ["src/text2video"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx",
    "Pillow",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

